+++ app/models/video.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+# Class for embedded videos
+class Video < ActiveRecord::Base
+  # Concerns
+  include Deletable
+
+  # Model Validation
+  validates :category_id, :name, :embed_url, presence: true
+
+  # Model Relationships
+  belongs_to :category
+
+  # Model Methods
+end